Create call-911.html
This video can be called up for any of the other pages/forms that exceeds safe diagnostic levels. The intent is eventually to add a user option of CLICK "YES"/"NO" if you want the app to dial 911 for you. Example - if patient's blood pressure is off the charts when they type in, this will give them the option to call 911. . .